Move single models to georef root dir
No point in having separate directories for single files
