+++ lib/assets/javascripts/cartodb/common/dialogs/georeference/admin_regions_model.js
@@ -1,7 +1,7 @@
 var cdb = require('cartodb.js');
 var Backbone = require('backbone');
-var StepsView = require('../steps_view');
-var RowModel = require('../row_model');
+var StepsView = require('./steps_view');
+var RowModel = require('./row_model');
 
 /**
  * Model for the administrative regions georeference option.
