+++ lib/assets/javascripts/cartodb/common/dialogs/georeference/georeference_model.js
@@ -1,11 +1,11 @@
 var Backbone = require('backbone');
 var _ = require('underscore');
 var cdb = require('cartodb.js');
-var LonLatColumnsModel = require('./lon_lat_columns/lon_lat_columns_model');
-var CityNamesModel = require('./city_names/city_names_model');
-var AdminRegionsModel = require('./admin_regions/admin_regions_model');
-var PostalCodesModel = require('./postal_codes/postal_codes_model');
-var IpAddressesModel = require('./ip_addresses/ip_addresses_model');
+var LonLatColumnsModel = require('./lon_lat_columns_model');
+var CityNamesModel = require('./city_names_model');
+var AdminRegionsModel = require('./admin_regions_model');
+var PostalCodesModel = require('./postal_codes_model');
+var IpAddressesModel = require('./ip_addresses_model');
 var StreetAddressesModel = require('./street_addresses/street_addresses_model');
 var GeocodeStuff = require('./geocode_stuff');
 var UserGeocodingModel = require('./user_geocoding_model');
