+++ lib/assets/javascripts/cartodb/common/dialogs/georeference/lon_lat_columns_model.js
@@ -1,7 +1,7 @@
 var cdb = require('cartodb.js');
 var Backbone = require('backbone');
-var RowModel = require('../row_model');
-var RowsView = require('../rows_view');
+var RowModel = require('./row_model');
+var RowsView = require('./rows_view');
 
 /**
  * Model for the Lon/Lat georeference option.
