+++ lib/assets/javascripts/cartodb/common/dialogs/georeference/postal_codes_model.js
@@ -1,18 +1,17 @@
-var _ = require('underscore');
 var cdb = require('cartodb.js');
 var Backbone = require('backbone');
-var StepsView = require('../steps_view');
-var RowModel = require('../row_model');
+var RowModel = require('./row_model');
+var StepsView = require('./steps_view');
 
 /**
- * Model for the city names georeference option.
+ * Model for the postal codes georeference option.
  */
 module.exports = cdb.core.Model.extend({
 
-  TAB_LABEL: 'City Names',
-  TITLE: "Select the column that contains the City's Name",
-  DESC: 'No matter the type of the columns you select, we will transform them to number for georeferencing.',
-  KIND: 'namedplace',
+  TAB_LABEL: 'Postal Codes',
+  TITLE: 'Select the column that has the Postal Codes',
+  DESC: 'Georeference your data by postal codes.',
+  KIND: 'postalcode',
 
   defaults: {
     step: 0,
@@ -75,18 +74,12 @@ module.exports = cdb.core.Model.extend({
 
   _geocode: function() {
     var d = this.get('geocodeStuff').geocodingChosenData({
-      type: 'city',
+      type: 'postal',
       kind: this.KIND,
       location: this._locationValue(),
       column_name: this._columnNameValue(),
       geometry_type: this.get('geometryType')
     }, this._isLocationFreeText());
-
-    var region = this._regionValue();
-    if (!_.isEmpty(region)) {
-      d.region = region;
-      d.region_text = this._isRegionFreeText();
-    }
 
     this.set('geocodeData', d);
   },
@@ -95,16 +88,12 @@ module.exports = cdb.core.Model.extend({
     var rows = new Backbone.Collection([
       new RowModel({
         comboViewClass: 'Combo',
-        label: 'In which column are your city names stored?',
+        label: 'In which column are your postal codes stored?',
         placeholder: 'Select column',
         data: this.get('columnsNames')
       }),
       new RowModel({
-        label: "Admin. Region where city's located, if known",
-        data: this.get('columns')
-      }),
-      new RowModel({
-        label: "Country where city's located, if known",
+        label: 'Country where postal codes are located, if known',
         data: this.get('columns')
       })
     ]);
@@ -113,18 +102,6 @@ module.exports = cdb.core.Model.extend({
 
   _columnNameValue: function() {
     return this.get('rows').first().get('value');
-  },
-
-  _regionValue: function() {
-    return this._region().get('value');
-  },
-
-  _isRegionFreeText: function() {
-    return this._region().get('isFreeText');
-  },
-
-  _region: function() {
-    return this.get('rows').at(1); // 2nd row
   },
 
   _locationValue: function() {
