+++ lib/assets/test/spec/cartodb/common/dialogs/georeference/admin_regions_model.spec.js
@@ -1,7 +1,7 @@
-var AdminRegionsModel = require('../../../../../../../javascripts/cartodb/common/dialogs/georeference/admin_regions/admin_regions_model');
-var GeocodeStuff = require('../../../../../../../javascripts/cartodb/common/dialogs/georeference/geocode_stuff');
+var AdminRegionsModel = require('../../../../../../javascripts/cartodb/common/dialogs/georeference/admin_regions_model');
+var GeocodeStuff = require('../../../../../../javascripts/cartodb/common/dialogs/georeference/geocode_stuff');
 
-describe('common/dialog/georeference/admin_regions/admin_regions_model', function() {
+describe('common/dialog/georeference/admin_regions_model', function() {
   beforeEach(function() {
     this.geocodeStuff = new GeocodeStuff('table_id');
     this.model = new AdminRegionsModel({
