+++ lib/assets/test/spec/cartodb/common/dialogs/georeference/ip_addresses_model.spec.js
@@ -1,7 +1,7 @@
-var IpAddressesModel = require('../../../../../../../javascripts/cartodb/common/dialogs/georeference/ip_addresses/ip_addresses_model');
-var GeocodeStuff = require('../../../../../../../javascripts/cartodb/common/dialogs/georeference/geocode_stuff');
+var IpAddressesModel = require('../../../../../../javascripts/cartodb/common/dialogs/georeference/ip_addresses_model');
+var GeocodeStuff = require('../../../../../../javascripts/cartodb/common/dialogs/georeference/geocode_stuff');
 
-describe('common/dialog/georeference/ip_addresses/ip_addresses_model', function() {
+describe('common/dialog/georeference/ip_addresses_model', function() {
   beforeEach(function() {
     this.geocodeStuff = new GeocodeStuff('table_id');
     this.model = new IpAddressesModel({
