+++ lib/assets/test/spec/cartodb/common/dialogs/georeference/lon_lat_columns_model.spec.js
@@ -1,8 +1,7 @@
-var cdb = require('cartodb.js');
-var LonLatColumnsModel = require('../../../../../../../javascripts/cartodb/common/dialogs/georeference/lon_lat_columns/lon_lat_columns_model');
-var GeocodeStuff = require('../../../../../../../javascripts/cartodb/common/dialogs/georeference/geocode_stuff');
+var LonLatColumnsModel = require('../../../../../../javascripts/cartodb/common/dialogs/georeference/lon_lat_columns_model');
+var GeocodeStuff = require('../../../../../../javascripts/cartodb/common/dialogs/georeference/geocode_stuff');
 
-describe('common/dialog/georeference/lon_lat_columns/lon_lat_columns_model', function() {
+describe('common/dialog/georeference/lon_lat_columns_model', function() {
   beforeEach(function() {
     this.geocodeStuff = new GeocodeStuff('table_id');
     this.model = new LonLatColumnsModel({
