+++ lib/assets/test/spec/cartodb/common/dialogs/georeference/postal_codes_model.spec.js
@@ -1,10 +1,10 @@
-var CityNamesModel = require('../../../../../../../javascripts/cartodb/common/dialogs/georeference/city_names/city_names_model');
-var GeocodeStuff = require('../../../../../../../javascripts/cartodb/common/dialogs/georeference/geocode_stuff');
+var PostalCodesModel = require('../../../../../../javascripts/cartodb/common/dialogs/georeference/postal_codes_model');
+var GeocodeStuff = require('../../../../../../javascripts/cartodb/common/dialogs/georeference/geocode_stuff');
 
-describe('common/dialog/georeference/city_names/city_names_model', function() {
+describe('common/dialog/georeference/postal_codes_model', function() {
   beforeEach(function() {
     this.geocodeStuff = new GeocodeStuff('table_id');
-    this.model = new CityNamesModel({
+    this.model = new PostalCodesModel({
       geocodeStuff: this.geocodeStuff,
       columnsNames: ['foo', 'lon', 'bar', 'lat'],
       columns: [
@@ -33,7 +33,7 @@ describe('common/dialog/georeference/city_names/city_names_model', function() {
 
     it('should call the geocode stuff model with expected args', function() {
       expect(this.geocodeStuff.availableGeometriesFetchData).toHaveBeenCalled();
-      expect(this.geocodeStuff.availableGeometriesFetchData.calls.argsFor(0)[0]).toEqual('namedplace');
+      expect(this.geocodeStuff.availableGeometriesFetchData.calls.argsFor(0)[0]).toEqual('postalcode');
       expect(this.geocodeStuff.availableGeometriesFetchData.calls.argsFor(0)[1]).toEqual('loc');
       expect(this.geocodeStuff.availableGeometriesFetchData.calls.argsFor(0)[2]).toEqual(true);
       expect(this.results).toBe(this.returnVal);
@@ -41,7 +41,7 @@ describe('common/dialog/georeference/city_names/city_names_model', function() {
   });
 
   describe('.continue', function() {
-    describe('when have selected a city column name', function() {
+    describe('when have selected a column name', function() {
       beforeEach(function() {
         this.model.set('canContinue', true);
         this.model.get('rows').first().set('value', 'col');
@@ -58,46 +58,21 @@ describe('common/dialog/georeference/city_names/city_names_model', function() {
         beforeEach(function() {
           this.model.set({
             canContinue: true,
-            geometryType: 'point'
+            geometryType: 'polygon'
           });
           this.model.get('rows').first().set('value', 'col');
-          this.model._location().set('value', 'loc');
+          this.model.get('rows').last().set('value', 'loc');
           this.model.continue();
         });
 
         it('should set geocode data', function() {
           var d = this.model.get('geocodeData');
           expect(d).toBeDefined();
-          expect(d.type).toEqual('city');
-          expect(d.kind).toEqual('namedplace');
+          expect(d.type).toEqual('postal');
+          expect(d.kind).toEqual('postalcode');
           expect(d.column_name).toEqual('col');
           expect(d.location).toEqual('loc');
-          expect(d.geometry_type).toEqual('point');
-        });
-
-        it('should not set geocode data with optional stuff', function() {
-          var d = this.model.get('geocodeData');
-          expect(d.region).toBeUndefined();
-          expect(d.region_text).toBeUndefined();
-        });
-
-        describe('when have also set a region', function() {
-          beforeEach(function() {
-            this.model._region().set('value', 'val');
-          });
-
-          it('should set geocode data with region', function() {
-            this.model.continue();
-            var d = this.model.get('geocodeData');
-            expect(d.region).toEqual('val');
-            expect(d.region_text).toBeFalsy();
-
-            this.model._region().set('isFreeText', true);
-            this.model.continue();
-            d = this.model.get('geocodeData');
-            expect(d.region).toEqual('val');
-            expect(d.region_text).toBe(true);
-          });
+          expect(d.geometry_type).toEqual('polygon');
         });
       });
     });
